add rule validator module
